addressed missed out comments in #1, except checkpointing
