--- src/instructlab/dolomite/hf_models/defaults.py
@@ -1,4 +1,0 @@
-# ----------------------------------------------------------------
-# Extracted from https://github.com/ibm-granite/dolomite-engine
-# ----------------------------------------------------------------
-DEFAULT_NORMALIZATION_IMPLEMENTATION = "torch"
