+++ src/instructlab/dolomite/hf_models/modeling_utils/attention/base.py
@@ -8,18 +8,18 @@ from typing import Tuple
 from transformers import DynamicCache
 import torch
 import torch.nn.functional as F
+from torch.nn import Linear # replaces ParameterizedLinear
 
 # Local
-from ...config import CommonConfig
+from ...models.gpt_dolomite.config import GPTDolomiteConfig
 from ...enums import AttentionHeadType, PositionEmbeddingType
-from ..linear import Linear
 from ..position_embedding import apply_rotary_pos_emb
 from .utils import repeat_key_value
 
 
 class Attention(torch.nn.Module):
     def __init__(
-        self, config: CommonConfig, causal: bool, layer_idx: int = None
+        self, config: GPTDolomiteConfig, causal: bool, layer_idx: int = None
     ) -> None:
         super().__init__()
