--- src/instructlab/dolomite/hf_models/modeling_utils/embedding.py
@@ -1,8 +1,0 @@
-# ----------------------------------------------------------------
-# Extracted from https://github.com/ibm-granite/dolomite-engine
-# ----------------------------------------------------------------
-# pylint: disable=unused-import
-# Third Party
-from torch.nn import Embedding
-
-# NOTE: we have replaced ParameterizedEmbedding with torch.nn.Embedding
