--- src/instructlab/dolomite/hf_models/modeling_utils/linear.py
@@ -1,8 +1,0 @@
-# ----------------------------------------------------------------
-# Extracted from https://github.com/ibm-granite/dolomite-engine
-# ----------------------------------------------------------------
-# pylint: disable=unused-import
-# Third Party
-from torch.nn import Linear
-
-# NOTE: we have replaced ParameterizedLinear with torch.nn.Linear
