--- src/instructlab/dolomite/hf_models/modeling_utils/normalization/rmsnorm/__init__.py
@@ -1,25 +1,0 @@
-# ----------------------------------------------------------------
-# Extracted from https://github.com/ibm-granite/dolomite-engine
-# ----------------------------------------------------------------
-# Third Party
-import torch
-
-# Local
-from .base import RMSNorm
-
-_RMSNORM_MODULES = {"torch": RMSNorm}
-
-
-def get_rmsnorm(
-    normalized_shape: int,
-    eps: float,
-    normalization_implementation: str = "torch",
-) -> torch.nn.LayerNorm:
-    if normalization_implementation in _RMSNORM_MODULES:
-        return _RMSNORM_MODULES[normalization_implementation](
-            normalized_shape=normalized_shape, eps=eps
-        )
-
-    raise ValueError(
-        f"unexpected `normalization_implementation` {normalization_implementation}"
-    )
