--- src/instructlab/dolomite/utils/wrapper.py
@@ -1,24 +1,0 @@
-# ----------------------------------------------------------------
-# Extracted from https://github.com/ibm-granite/dolomite-engine
-# ----------------------------------------------------------------
-# Standard
-from typing import List, Type
-
-# Third Party
-import torch
-
-
-def get_module_class_from_name(
-    model: torch.nn.Module, name: str
-) -> List[Type[torch.nn.Module]]:
-    modules_children = list(model.children())
-
-    if model.__class__.__name__ == name:
-        return model.__class__
-    elif len(modules_children) == 0:
-        return
-    else:
-        for child_module in modules_children:
-            module_class = get_module_class_from_name(child_module, name)
-            if module_class is not None:
-                return module_class
